requirements.txt
openai
pydantic
instructor
Flask
Flask-SocketIO
python-dotenv
networkx
Faker
tavily-python
